♻️ F10: BASKET > Add useMenu.jsx
