+++ src/components/pages/Order/OrderPage.jsx
@@ -5,65 +5,21 @@ import { styled } from "styled-components";
 import { theme } from "../../..";
 import { useRef, useState } from "react";
 import OrderContext from "../../../context/OrderContext";
-import { fakeMenu } from "../../../../data/fakeMenu";
 import { EMPTY_PRODUCT } from "../../../enums/products";
-import { deepClone } from "../../../utils/array";
+import { useMenu } from "../../../hooks/useMenu";
 
 export default function OrderPage() {
  // state
  const [isModeAdmin, setIsModeAdmin] = useState(true)
  const [isCollapsed, setIsCollapsed] = useState(false)
  const [currentTabSelected, setCurrentTabSelected] = useState("edit")
- const [menu, setMenu] = useState(fakeMenu.SMALL)
  const [newProduct, setNewProduct] = useState(EMPTY_PRODUCT);
- const [productSelected, setProductSelected] = useState(fakeMenu.SMALL[0]);
+ const [productSelected, setProductSelected] = useState(EMPTY_PRODUCT);
  const titleEditRef = useRef();
-
+const {menu, setMenu, handleAdd, handleDelete, resetMenu, handleEdit} = useMenu()
 
 
  //CRUD
-//comportements (gestionnaire de state ou "state handlersg")
-const handleAdd = (newProduct) => {
-
-  // 1. copie du tableau
-  const menuCopy = deepClone(menu)
-
-  // 2. ajout du produit
-  const menuUpdated = [ newProduct, ...menuCopy ]
-
-  // 3. mise à jour du state
-  setMenu(menuUpdated)
-}
-
-const handleDelete = (idOfProductToDelete) => {
-  //copy du state
-  const menuCopy = deepClone(menu)
-  //manipulation du state
-  const menuUpdated = menuCopy.filter((product) => product.id !== idOfProductToDelete);
-  // console.log("menuUpdated: ", menuUpdated);
-  //mise à jour du state
-  setMenu(menuUpdated);
-};
-const handleEdit = (productBeingEdited) => { 
-
-console.log("productBeingEdited: ", productBeingEdited);
-  //1. Copie du state (deep clone)
-  const menuCopy = deepClone(menu)
-
-  //2. Manipulation de la copie du state
-  const indexOfProductToEdit = menu.findIndex(
-    (menuProduct) => menuProduct.id === productBeingEdited.id
-    )
-  // console.log("indexOfProductToEdit: ", indexOfProductToEdit);
-// }
-  menuCopy[indexOfProductToEdit] = productBeingEdited
-  //3. Mise à jour du state
-  setMenu(menuCopy)
-}
-
-const resetMenu = () => {
-  setMenu(fakeMenu.MEDIUM);
-}; 
 
 const orderContextValue = {
   isModeAdmin,
